requirements.txt
fastapi
uvicorn
cohere
faiss-cpu
numpy
requests
